Word audio generation via OpenAI
