+++ lunes_cms/cmsv2/admins/word_admin.py
@@ -1,7 +1,8 @@
 from __future__ import absolute_import, unicode_literals
 
 from django.contrib import admin
-from django.utils.html import mark_safe
+from django.urls import reverse
+from django.utils.html import mark_safe, format_html
 from django.utils.translation import gettext_lazy as _
 
 from lunes_cms.cmsv2.admins.base import BaseAdmin
@@ -37,6 +38,8 @@ class WordAdmin(BaseAdmin):
     It includes custom display methods for showing and managing assets.
     """
 
+    audio_generate = ""
+
     fields = (
         "word_type",
         "grammatical_gender",
@@ -45,6 +48,7 @@ class WordAdmin(BaseAdmin):
         "plural_article",
         "plural",
         "audio",
+        "audio_generate",
         "audio_check_status",
         "image",
         "image_check_status",
@@ -53,7 +57,7 @@ class WordAdmin(BaseAdmin):
         "additional_meaning_1",
         "additional_meaning_2",
     )
-    readonly_fields = ("created_by", "image_tag")
+    readonly_fields = ("audio_generate", "created_by", "image_tag")
     search_fields = ["word"]
     ordering = ["word", "creation_date"]
     inlines = [UnitInline]
@@ -88,6 +92,16 @@ class WordAdmin(BaseAdmin):
             "js/image_check_status_update.js",
         ]
         css = {"all": ["css/asset_manager.css", "css/audio_player.css"]}
+
+    def audio_generate(self, obj):
+        if obj.pk:
+            url = reverse("cmsv2:word_generate_audio", args=[obj.pk])
+            return format_html(
+                '<a class="button" href="{}">Generate Audio</a>', url
+            )
+        return "Save to enable audio generation."
+
+    audio_generate.short_description = "Audio Generation"
 
     def creator_group(self, obj):
         """
@@ -163,12 +177,13 @@ class WordAdmin(BaseAdmin):
             selected = "selected" if obj.audio_check_status == value else ""
             options += f'<option value="{value}" {selected}>{display}</option>'
 
-        html = f"""
-        {word_audio_container}
-        <select name="audio_check_status_{obj.id}" data-word-id="{obj.id}" class="audio-check-status-select" style="margin-top: 8px;">
-            {options}
-        </select>
-        """
+        html = word_audio_container
+        if obj.audio:
+            html += f"""
+            <select name="audio_check_status_{obj.id}" data-word-id="{obj.id}" class="audio-check-status-select" style="margin-top: 8px;">
+                {options}
+            </select>
+            """
 
         return mark_safe(html)
 
@@ -237,7 +252,11 @@ class WordAdmin(BaseAdmin):
         </select>
         """
 
-        return f'<div class="word-image-container">{image_html}{controls_html}</div>{word_image_check_status_html}'
+        html = f'<div class="word-image-container">{image_html}{controls_html}</div>'
+        if obj.image:
+            html += word_image_check_status_html
+
+        return html
 
     def _generate_unit_word_images(self, obj):
         """
@@ -297,16 +316,20 @@ class WordAdmin(BaseAdmin):
 
         unit_name_html = f'<div class="unit-name">{unit_name}</div>'
 
-        return f"""
+        html = f"""
         <div class="unitword-image-wrapper">
             {unit_name_html}
             <div class="unitword-image-container">
                 {unit_image_html}
                 {unit_controls_html}
             </div>
-            {unit_image_check_status_html}
         </div>
         """
+
+        if relation.image:
+            html += unit_image_check_status_html
+
+        return html
 
     list_image.short_description = _("Image")
