+++ setup.cfg
@@ -52,6 +52,7 @@ install_requires =
 	djangorestframework==3.15.2
 	drf-spectacular==0.27.2
 	ipython==8.29.0
+	openai==1.88.0
 	pillow==11.0.0
 	psycopg2==2.9.10
 	pydub==0.25.1
